+++ packages/@expo/prebuild-config/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### 🎉 New features
 
 - Add android config plugin for app name translation. ([#37202](https://github.com/expo/expo/pull/37202) by [@aleqsio](https://github.com/aleqsio))
+- Support Liquid Glass app icons.
 
 ### 🐛 Bug fixes